get rid of duplicate stripe error object
