+++ support-workers/src/main/scala/com/gu/stripe/StripeService.scala
@@ -67,29 +67,6 @@ class StripeServiceForCurrency(val stripeService: StripeService, currency: Curre
 
 }
 
-object StripeError {
-  private val encoder = deriveEncoder[StripeError].mapJson { json => Json.fromFields(List("error" -> json)) }
 
-  private val decoder = deriveDecoder[StripeError].prepare { _.downField("error") }
 
-  implicit val codec: Codec[StripeError] = new Codec[StripeError](encoder, decoder)
-}
 
-//See docs here: https://stripe.com/docs/api/curl#errors
-case class StripeError(
-  `type`: String, //The type of error: api_connection_error, api_error, authentication_error, card_error, invalid_request_error, or rate_limit_error
-  message: String, //A human-readable message providing more details about the error. For card errors, these messages can be shown to your users.
-  code: Option[String] = None, //For card errors, a short string from amongst those listed on the right describing the kind of card error that occurred.
-  decline_code: Option[String] = None, //For card errors resulting from a bank decline, a short string indicating the bank's reason for the decline.
-  param: Option[String] = None //The parameter the error relates to if the error is parameter-specific..
-) extends Throwable {
-
-  override def getMessage: String =
-    s"message: $message; type: ${`type`}; code: ${code.getOrElse("")}; decline_code: ${decline_code.getOrElse("")}; param: ${param.getOrElse("")}"
-
-  def asRetryException: RetryException = `type` match {
-    case ("api_connection_error" | "api_error" | "rate_limit_error") => new RetryUnlimited(this.asJson.noSpaces, cause = this)
-    case "authentication_error" => new RetryLimited(this.asJson.noSpaces, cause = this)
-    case ("card_error" | "invalid_request_error" | "validation_error") => new RetryNone(this.asJson.noSpaces, cause = this)
-  }
-}
