+++ support-workers/src/test/scala/com/gu/support/workers/errors/StripeErrorsSpec.scala
@@ -11,6 +11,7 @@ import com.gu.support.workers.JsonFixtures.{createStripeSourcePaymentMethodContr
 import com.gu.support.workers.exceptions.{RetryNone, RetryUnlimited}
 import com.gu.support.workers.lambdas.CreatePaymentMethod
 import com.gu.support.workers.{AsyncLambdaSpec, JsonFixtures, JsonWrapper, MockContext}
+import io.circe.Json
 import io.circe.parser.decode
 import io.circe.syntax._
 
@@ -47,6 +48,9 @@ class StripeErrorsSpec extends AsyncLambdaSpec with MockWebServerCreator with Mo
   }
 
   "402s from Stripe" should "throw a FatalException" in {
+    import io.circe.generic.semiauto.deriveEncoder
+    implicit val encoder = deriveEncoder[StripeError].mapJson { json => Json.fromFields(List("error" -> json)) }
+
     val errorJson = StripeError("card_error", "The card has expired", Some("expired_card")).asJson.noSpaces
     val server = createMockServer(402, errorJson)
     val baseUrl = server.url("/v1")
